refactor: move subscriptions into execution folder
All operations, including subscription operations, are covered by the execution section of the GraphQL spec
