+++ src/README.md
@@ -20,4 +20,3 @@ Each sub directory within is a sub-module of graphql-js:
 - [`graphql/error`](error/README.md): Creating and formatting GraphQL errors.
 - [`graphql/utilities`](utilities/README.md): Common useful computations upon
   the GraphQL language and type objects.
-- [`graphql/subscription`](subscription/README.md): Subscribe to data updates.
