+++ src/execution/README.md
@@ -7,3 +7,8 @@ fulfilling a GraphQL request.
 import { execute } from 'graphql/execution'; // ES6
 var GraphQLExecution = require('graphql/execution'); // CommonJS
 ```
+
+```js
+import { subscribe, createSourceEventStream } from 'graphql/execution'; // ES6
+var GraphQLSubscription = require('graphql/execution'); // CommonJS
+```
