--- src/subscription/index.ts
@@ -1,2 +1,0 @@
-export { subscribe, createSourceEventStream } from './subscribe';
-export type { SubscriptionArgs } from './subscribe';
